game.java
his Tic Tac Toe Game is a Graphical user interface(GUI) Game that allows user to easily Play the game using buttons. Users may very quickly play the Game using the system simple design. The Tic Tac Toe Game has simple design with buttons. The player can click on the button one by one to play the game.

The users may play the game using buttons. This game can be played by 2 users. The 1st player can click on any button the value of the button will be ‘x’.  When the 2nd player clicks the button the value of button will be ‘o’.  The winner of the game is decided based on the rows, column or diagonal values should be same.

The Java Swing library which provides flexible set of components is used to build the Tic Tac Toe Game. The application makes use of Swing components such as Jbutton, JPanel and JFrame among many other Swing components. Button event is handled by the ActionListener interface and button events are defined using ActionListener’s actionPerformed() method. The winner of the game is checked using checkForWinner() method.

Methods used in the project:

actionPerformed():

The action or events of the buttons are handled using this method. Buttons in the system to generate the value of button is handled using this method. This method describes how the component should behave after the button is click.

checkForWinner():

This is a user defined method to check the winner of game. To win the game the any one player needs the value of row, column or diagonal should be same. If values of row, column or diagonal is not same it pop ups the dialog box with ‘tie’ message.

resetGame():

This is a user defined method to reset the game when the winner of game is declared. The value of buttons are assigned to null when this method is triggered.